Added a command-line option for stripping pauses from presentations (which are useful for getting handouts).
Moreover, there's a fix for relative image paths expressed either as markdown images (i.e., with the ![]() syntac) or as html images (i.e., with the <img> tag).
In particular, the relative image paths in case of export and pdf are resolved to the absolute path in order to properly have the image included in the output.