Split and reassamble markdown at </script>
After JSON.stringifying the whole md, it is cut in chunks at the / in
each </script>, and reassambled as sum of strings. For example if the
resulting string would be:

    "This contains <script>foo</script> tags"

It would be converted to:

    "This cointains <script>foo<" + "/script> tags"

The resulting string is the same, but being cut this way the browser is
no longer confused by the `</script>` inside the code.