Better handling of paths
When exporting to pdf/.tmp and *not* inlining large assets (> 8kb as data url),
Chrome is still not going to find these relative paths. So configure
`.css()` to rewrite file references to be relative to the export location,
and set relativeTo to the export directory so the assets are still found.